wip refactor injection & resolution
