+++ src/tests/core/test_agrupation.py
@@ -1,6 +1,7 @@
 from pydantic import BaseModel
-from dependency.core.agrupation import Plugin, PluginMeta, module
-from dependency.core.injection import Container
+from dependency.core2.agrupation.module import module
+from dependency.core2.agrupation.plugin import Plugin, PluginMeta
+from dependency.core2.container import Container
 
 class TPluginConfig(BaseModel):
     field1: str
