+++ src/tests/core/test_injection.py
@@ -1,6 +1,8 @@
 from dependency_injector import containers, providers
 from dependency_injector.wiring import Provide, inject
-from dependency.core.injection import ContainerInjection, ProviderInjection
+from dependency.core2.injection.base import ContainerInjection
+from dependency.core2.injection.provider import ProviderInjection
+from dependency.core2.resolution.implementation import Implementation
 
 TEST_REFERENCE = "container1.container2.provider1"
 
@@ -18,13 +20,12 @@ def test_injection1():
     container2 = ContainerInjection(name="container2", parent=container1)
     provider1 = ProviderInjection(
         name="provider1",
-        component_name="provider1",
-        interface_cls=Interface,
         parent=container2)
-    provider1.set_implementation(
+    provider1.implementation = Implementation(
         provided_cls=Instance,
         provider_cls=providers.Singleton,
-        component_cls=Interface)
+        component_cls=Interface
+    )
     assert provider1.reference == TEST_REFERENCE
 
     container = containers.DynamicContainer()
